requirements.txt
djangorestframework
drf-yasg
django-ckeditor
django-cors-headers
mysqlclient
pymysql
django-jazzmin
